+++ main.go
@@ -2,8 +2,8 @@ package main
 
 import (
 	"fmt"
-	csvmanager "github.com/georgexx009-wizeline/ondemand-go-bootcamp/utils/csv-manager"
-	"github.com/georgexx009-wizeline/ondemand-go-bootcamp/utils/fetcher"
+	csvmanager "github.com/georgexx009-wizeline/ondemand-go-bootcamp/pkg/csv-manager"
+	"github.com/georgexx009-wizeline/ondemand-go-bootcamp/pkg/fetcher"
 
 	"github.com/georgexx009-wizeline/ondemand-go-bootcamp/controller"
 	"github.com/georgexx009-wizeline/ondemand-go-bootcamp/router"
